환전기능 적용
