+++ client/src/components/Router.js
@@ -5,7 +5,6 @@ import PiggyBoss from "../Routes/PiggyBoss";
 import AdminPage from "../Routes/AdminPage";
 import LoginForm from "../Routes/LoginForm";
 import EmailAuth from "../Routes/EmailAuth";
-import Currency from "../Routes/Currency";
 
 function AppRouter() {
   return (
@@ -15,7 +14,6 @@ function AppRouter() {
         <Route path="/register" element={<EmailAuth />}></Route>
         <Route path="/piggy" element={<PiggyBoss />}></Route>
         <Route path="/admin" element={<AdminPage />}></Route>
-        <Route path="/currency" element={<Currency />}></Route>
       </Routes>
     </Router>
   );
